made huge refactors on all components so it shares the same componenet for recipes and blogs. Added initial hookup to search bar and autocomplete onchange handlers
